Improvements to test coverage and reduced test times
