+++ pkg/goSignals/server/test/server_test.go
@@ -9,73 +9,25 @@ import (
 	"encoding/pem"
 	"fmt"
 	"io"
-	"log"
 	"net/http"
-	"net/http/httptest"
 	"net/url"
-	"os"
 	"strings"
 	"testing"
 	"time"
 
+	"github.com/MicahParks/keyfunc/v2"
+	"github.com/golang-jwt/jwt/v5"
 	"github.com/i2-open/i2goSignals/internal/authUtil"
-	"github.com/i2-open/i2goSignals/internal/providers/dbProviders"
-	"github.com/i2-open/i2goSignals/pkg/authSupport"
 	"github.com/i2-open/i2goSignals/pkg/constants"
 	"github.com/i2-open/i2goSignals/pkg/goSet"
 	ssef "github.com/i2-open/i2goSignals/pkg/goSignals/server"
 	"github.com/i2-open/i2goSignals/pkg/ssfModels"
-	"github.com/i2-open/i2goSignals/pkg/tlsSupport"
-	"go.mongodb.org/mongo-driver/v2/bson"
-
-	"github.com/MicahParks/keyfunc/v2"
-	"github.com/golang-jwt/jwt/v5"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/testutil"
 )
-
-var TestDbUrl = "mongodb://root:dockTest@mongo1:30001,mongo2:30002,mongo3:30003/?retryWrites=true&replicaSet=dbrs&readPreference=primary&serverSelectionTimeoutMS=5000&connectTimeoutMS=10000&authSource=admin&authMechanism=SCRAM-SHA-256"
-
-var testLog = log.New(os.Stdout, "TEST: ", log.Ldate|log.Ltime)
-
-type ssfInstance struct {
-	ts              *httptest.Server
-	host            string
-	client          *http.Client
-	provider        dbProviders.DbProviderInterface
-	stream          model.StreamConfiguration
-	app             *ssef.SignalsApplication
-	streamToken     string
-	streamMgmtToken string
-	iatToken        string
-	projectId       string
-	startTime       *time.Time
-}
-
-func (instance *ssfInstance) GetPollUrl(stream model.StreamConfiguration) string {
-	if stream.Delivery == nil || stream.Delivery.PollTransmitMethod == nil {
-		return ""
-	}
-	endpoint := stream.Delivery.PollTransmitMethod.EndpointUrl
-	if strings.HasPrefix(endpoint, "http") {
-		return endpoint
-	}
-	return instance.ts.URL + endpoint
-}
-
-func (instance *ssfInstance) GetPushUrl(stream model.StreamConfiguration) string {
-	if stream.Delivery == nil || stream.Delivery.PushReceiveMethod == nil {
-		return ""
-	}
-	endpoint := stream.Delivery.PushReceiveMethod.EndpointUrl
-	if strings.HasPrefix(endpoint, "http") {
-		return endpoint
-	}
-	return instance.ts.URL + endpoint
-}
 
 type ServerSuite struct {
 	suite.Suite
@@ -124,72 +76,7 @@ func TestServer(t *testing.T) {
 	testLog.Println("** TEST COMPLETE **")
 }
 
-func createServer(t *testing.T, dbName string, resetDb bool) (*ssfInstance, error) {
-	t.Helper()
-	var err error
-	var instance ssfInstance
-
-	dbUrl := "memorydb:"
-	if os.Getenv("TEST_MONGO_CLUSTER") != "" {
-		dbUrl = TestDbUrl
-	} else {
-		// When using memory provider, use a temporary directory for persistence
-		// to avoid leaving files in the source tree.
-		t.Setenv("MEM_DIRECTORY", t.TempDir())
-	}
-	// mongo, err := mongo_provider.Open(TestDbUrl, dbName)
-	mongo, err := dbProviders.OpenProvider(dbUrl, dbName)
-	if err != nil {
-		t.Error("Mongo client error: " + err.Error())
-		return nil, err
-	}
-
-	if resetDb {
-		_ = mongo.ResetDb(true)
-	}
-
-	// Build application and wrap with httptest.Server
-	app := ssef.NewApplication(mongo, "")
-	ts := httptest.NewServer(app.Handler)
-	instance.ts = ts
-	instance.app = app
-	u, _ := url.Parse(ts.URL)
-	instance.host = u.Host
-	// Set BaseUrl on app for any logic that depends on it
-	baseUrl, _ := url.Parse(ts.URL + "/")
-	app.SetBaseUrl(baseUrl)
-	instance.client = ts.Client()
-	tlsSupport.CheckCaInstalled(instance.client)
-	instance.provider = mongo
-	nowTime := time.Now()
-	instance.startTime = &nowTime
-
-	instance.iatToken, err = instance.provider.GetAuthIssuer().IssueProjectIat(nil)
-	if err != nil {
-		t.Logf("Error creating iat: %s\n", err.Error())
-	}
-	eat, err := instance.provider.GetAuthIssuer().ParseAuthToken(instance.iatToken)
-	if err != nil {
-		t.Logf("Error parsing iat: %s\n", err.Error())
-	}
-
-	clientToken, err := instance.provider.GetAuthIssuer().IssueStreamClientToken(model.SsfClient{
-		Id:            bson.ObjectID{},
-		ProjectIds:    []string{eat.ProjectId},
-		AllowedScopes: []string{authSupport.ScopeStreamAdmin, authSupport.ScopeStreamMgmt},
-		Email:         "test@test.com",
-		Description:   "server test",
-	}, eat.ProjectId, true)
-	instance.streamMgmtToken = clientToken
-
-	instance.projectId = eat.ProjectId
-
-	return &instance, nil
-}
-
 func (suite *ServerSuite) TearDownTest() {
-	// Wait for tests to settle (database updates to complete
-	time.Sleep(500 * time.Millisecond)
 }
 
 // Test1_Certificate loads the servers default certificate in a couple of ways and attempts to parse it.
